Separate TaskLabel components depending on task type, fix issue with calendar button not closing calendar popover properly
